More Test Organisation
A few tests to add then test driving can continue
