+++ DivePlannerTests/ApplicationLayerTests/DivePlan/DiveModelUserInterfaceShould.cs
@@ -53,16 +53,5 @@ namespace DivePlannerTests
 
             //Assert
         }
-
-        //TODO Setup cannot execute in the view model command and test that
-        [Fact(Skip = "Test needs implementing - is can execute dive step false when selected dive model or gas mixture is null")]
-        public void NotAllowDivesProfilesToBeRunWithoutASelectedDiveModel()
-        {
-            //Arrange
-
-            //Act
-
-            //Assert
-        }
     }
 }
