+++ DivePlannerTests/ApplicationLayerTests/DiveResults/DiveWarningsUserInterfaceShould.cs
@@ -4,6 +4,20 @@ namespace DivePlannerTests
 {
     public class DiveWarningsUserInterfaceShould
     {
-       
+        [Fact(Skip = "Needs Implementing")]
+        public void PopulateDiveWarningsInResults()
+        {
+
+        }
+
+        [Fact(Skip = "Tests Needs implementing")]
+        public void RaisePropertyChangeWhenDiveWarningsIsPopulated()
+        {
+            //Arrange
+
+            //Act
+
+            //Assert
+        }
     }
 }
