+++ DivePlannerTests/ApplicationLayerTests/Header/LoadDiveUserInterfaceShould.cs
@@ -4,6 +4,10 @@ namespace DivePlannerTests
 {
     public class LoadDiveUserInterfaceShould
     {
-    
+        [Fact(Skip="Needs Implementing")]
+        public void LoadAnOpenJsonFileNativeDialog()
+        {
+
+        }
     }
 }
