+++ DivePlannerTests/ApplicationLayerTests/Header/PrintUserInterfaceShould.cs
@@ -4,6 +4,10 @@ namespace DivePlannerTests
 {
     public class PrintUserInterfaceShould
     {
-    
+        [Fact(Skip="Needs Implementing")]
+        public void OpenAPrintFileNativeDialog()
+        {
+
+        }
     }
 }
