+++ DivePlannerTests/ApplicationLayerTests/Header/RedoUserInterfaceShould.cs
@@ -4,6 +4,10 @@ namespace DivePlannerTests
 {
     public class RedoUserInterfaceShould
     {
-    
+        [Fact(Skip="Needs Implementing")]
+        public void RedoThePreviousDiveResultsStep()
+        {
+
+        }
     }
 }
