+++ DivePlannerTests/ApplicationLayerTests/Header/SaveDiveUserInterfaceShould.cs
@@ -4,6 +4,10 @@ namespace DivePlannerTests
 {
     public class SaveDiveUserInterfaceShould
     {
-    
+        [Fact(Skip="Needs Implementing")]
+        public void LoadASaveJsonFileNativeDialog()
+        {
+
+        }
     }
 }
