+++ DivePlannerTests/ApplicationLayerTests/Header/UndoUserInterfaceShould.cs
@@ -4,6 +4,10 @@ namespace DivePlannerTests
 {
     public class UndoUserInterfaceShould
     {
-    
+        [Fact(Skip="Needs Implementing")]
+        public void UndoTheCurrentDiveResultsStep()
+        {
+
+        }
     }
 }
