+++ DivePlannerTests/BusinessLayerTests/Header/LoadDiveBusinessLayerShould.cs
@@ -1,0 +1,25 @@
+using Xunit;
+
+namespace DivePlannerTests
+{
+    public class LoadDiveBusinessLayerShould
+    {
+         [Fact(Skip="Needs Implementing")]
+        public void ValidateTheJsonFileContent()
+        {
+
+        }
+
+        [Fact(Skip="Needs Implementing")]
+        public void LoadDiveFromJsonFileToEntityModel()
+        {
+
+        }
+
+        [Fact(Skip="Needs Implementing")]
+        public void ConvertFromEntityModelsToViewModels()
+        {
+
+        }
+    }
+}
