+++ DivePlannerTests/BusinessLayerTests/Header/PrintBusinessLayerShould.cs
@@ -1,0 +1,25 @@
+using Xunit;
+
+namespace DivePlannerTests
+{
+    public class PrintBusinessLayerShould
+    {
+        [Fact(Skip="Needs Implementing")]
+        public void PopulateATemporaryJsonFileFromTheDiveResults()
+        {
+
+        }
+
+        [Fact(Skip="Needs Implementing")]
+        public void PrintTheTemporaryJsonFile()
+        {
+
+        }
+
+        [Fact(Skip="Needs Implementing")]
+        public void DeleteTheTemporaryJsonFile()
+        {
+
+        }
+    }
+}
