+++ DivePlannerTests/BusinessLayerTests/Header/RedoBusinessLayerShould.cs
@@ -1,0 +1,19 @@
+using Xunit;
+
+namespace DivePlannerTests
+{
+    public class RedoBusinessLayerShould
+    {
+        [Fact(Skip="Needs Implementing")]
+        public void ReinstateThePreviouslyRemovedDiveResultsStep()
+        {
+
+        }
+
+        [Fact(Skip="Needs Implementing")]
+        public void RollBackTheRevertToTheDiveServiceOfTheLastDiveStepState()
+        {
+
+        }
+    }
+}
