+++ DivePlannerTests/BusinessLayerTests/Header/SaveDiveBusinessLayerShould.cs
@@ -1,0 +1,19 @@
+using Xunit;
+
+namespace DivePlannerTests
+{
+    public class SaveDiveBusinessLayerShould
+    {
+        [Fact(Skip="Needs Implementing")]
+        public void ConvertFromViewModelsToEntityModels()
+        {
+
+        }
+
+        [Fact(Skip="Needs Implementing")]
+        public void SaveDiveToJsonFile()
+        {
+
+        }
+    }
+}
