+++ DivePlannerTests/BusinessLayerTests/Header/UndoBusinessLayerShould.cs
@@ -1,0 +1,19 @@
+using Xunit;
+
+namespace DivePlannerTests
+{
+    public class UndoBusinessLayerShould
+    {
+        [Fact(Skip="Needs Implementing")]
+        public void RemoveTheLatestCurrentDiveResultsStep()
+        {
+
+        }
+
+        [Fact(Skip="Needs Implementing")]
+        public void RevertTheDiveServiceBackFromTheLastDiveStepState()
+        {
+
+        }
+    }
+}
